move logic from DatabaseHelper to TasksQueryHelper
