+++ opencorpora/src/main/java/org/opencorpora/data/dal/TypesQueryHelper.java
@@ -1,0 +1,5 @@
+package org.opencorpora.data.dal;
+
+
+public class TypesQueryHelper {
+}
